requirements.txt
Django>=2.2.6
djangorestframework>=3.10.3
coreapi>=2.3.3
Markdown>=3.1.1
Pygments>=2.4.2